support for streams
I am using gnump3d to get my tracks to my laptop. I noticed that none
of the streamed tracks were getting to last.fm. I investigated and
discovered that mpdscribble was not getting told of the track length
by mpd -- understandably so -- and was not reporting it to last.fm.

This patch is more of a hack then a fix, but it works for me. :)