file: parse command line arguments before config file
Don't split the command line parser into two parts ("--conf", and the
rest).  Parse all of it at once, and open the configuration file after
that.  This requires a change in the configuration file parser:
existing values must not be overwritten, because the command line
should override the configuration file (being more explicit).