Don't assume last.fm AS_HOST + config parsing
In fact, don't assume global state for username/password either.
A lot of changes to as.c, conn.c, conn.h just to allow a pointer to a struct
containing the state instead of assuming global.

This patch still just uses the first as_host from the list.

file.h and file.c have been modified to actually parse the new config
file format and create a linked list of as_hosts.

There is a memory leak in the config parser: the linked list is not freed.
Since we need it all in memory until shutting down, this is likely not a problem,
but a comment reminder has been added.