system/Error: replace individual includes with Windows.h
You could probably pick out the required parts needed by each header to only include the parts you need. However, doing it this way causes way less headaches with the compiler and linker. Maybe in the future we can figure out a way to only pull out the parts we need and discard everything else, but relying on the compiler to properly tree shake is probably a better idea for now.